Added `icon` helper.
Adds a level of abstraction to the existing `glyphicon` helper, allowing
users to select a different vector icons library. For instance, it is
now possible to display Font Awesome icons by passing the library option:

    <%= icon :heart, library: :font_awesome %>

Note that, since Font Awesome CSS is not loaded by default with Bootstrap
CSS, it's up to the user to make sure the CSS is loaded in the same app.