Add partial support for Middleman
Ideally, Bh would not only work with Rails, but with any Ruby framework
that handle views, like Middleman and Padrino.

As of now, some helpers already work with Middleman (Cdn, Icon, Glyphicon),
while most of them do not, due to the fact that Padrino and ActionView implement
the same helpers in different ways (e.g., `link_to`, `form_for`).

This commit adds support to Middleman, although for now only the helpers
above will work. Once all helpers are supported, Middleman will be added
to the README and documentation.

This commit also adds a spec/dummy site running Middleman, to make it easier
for developers to see how the HTML generated by the helpers looks like in a
browser. Once again, this site only includes a few helpers for now, but the
plan is to include all of them.