Pass `progress_bar` extra params to the container
Before this commit, the `progress_bar` helper did not accept options
for the outer container. For instance the code:

```rhtml
<%= progress_bar({percentage: 30, label: 'With container params'}, class: :important, id: 'my-container') %>
```

would fail. After this commit, the result is:

```html
<div class="important progress" id="my-container">
  <div class="progress-bar" style="width: 30%" role="progressbar" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100">
    With container params
  </div>
</div>
```

which should partially mitigate #39.

Note that the parentheses and brackets are required to distinguish between
parameters passed to the bar and to the wrapping container.