Pass `id` param to navbar subcomponents
In a navbar, the ID of the horizontal element must match the 'data-target'
of the vertical element in order for the the navbar "toggle" to work.

Before this commit, a random ID was generated for each navbar.
After this commit, a specific ID can be assigned by passing the `:id`
parameter to the wrapping `navbar`.

Any other extra param of the `vertical` and `horizontal` helpers is passed
to their respective <div> elements, which should partially mitigate #39.