Make all non-Rails helpers work in Middleman too
Before this commit, non-Rails-specific Helpers were only working in
Middleman when called without a block; otherwise they would fail.

This commit completes the separation of Rails-specific helpers into
their own namespace which requires `action_view` helpers.

In this way, Middleman app can rely to their own Padrino helpers for
methods like `capture`, and can work even when passed a block.

Note that the Embedded Ruby syntax of Rails is slightly different from
the default when accepting a block. The default way (used by Middleman) is:

```rhtml
<% alert_box context: :success do %>
  Text <em>styled</em> with <%= link_to 'links', '#', class: 'alert-link' %>.
<% end %>
```

that is, starting with `<%`. Rails uses a different format:

```rhtml
<%= alert_box context: :success do %>
  Text <em>styled</em> with <%= link_to 'links', '#', class: 'alert-link' %>.
<% end %>
```
that is, starting with `<%=`.