Pass `id` param to dropdown subcomponents
In a dropdown, the ID of the button match the 'data-target' of the `<ul>`
element in order for the the dropdown "toggle" to work.

Before this commit, a random ID was generated for each dropdown.
After this commit, a specific ID can be assigned by passing the `:id`
parameter to the `dropdown` helper.