Bugfix: don’t *always* override `button_to`
Bh is meant to not override any existing helper by default.

For instance, `form_for` is only used if a `:layout` options is passed,
and `link_to` if used inside a different Bh helper.

This commit makes `button_to` override the original method *only* if at
least one of the options `:context`, `:size` or `:layout` is provided.

This makes sure that including `bh` in an app does not change any existing
behavior unless instructed to do so.

In particular, it prevents `class="btn"` to be appended to every button
created with `button_to`.