Don’t assume that the content of `link_to` is safe
Closes #138 – Security fix

Before this commit, `link_to '<script>alert("xss")</script>', '/'`
would actually execute the Javascript in the browser, resulting in
a security issue (XSS).

After this commit, Bh does not always assume that the content of
`link_to` is HTML-safe, but correctly lets the super method (of
Rails of Padrino) deal with this.